pyproject.toml
[tool.poetry]
name = "voice-doct-appointment"
version = "0.1.0"
description = "A voice-enabled doctor booking assistant"
authors = ["Dokeun Oh <ohdoking@gmail.com>"]
readme = "README.md"
packages = [
    { include = "voice_doctor_appointment", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.13.4"
requests = "^2.32.4"
elevenlabs = "^2.7.1"
sounddevice = "^0.5.2"
scipy = "^1.16.0"
openai = "^1.97.0"
streamlit = "^1.47.0"
numpy = "^2.3.1"
python-dotenv = "^1.0.0"
dataclasses-json = "^0.6.7"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
doctor-booking = "voice_doctor_appointment.app.app:main"
